pyproject.toml
[project]
name = "multi-agentic-ai-workshop"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[google,mcp]>=1.0.1",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "python-dotenv==1.1.1",
    "ruff>=0.12.12",
]
